changed field type for RequestedHours

+++ mslcschedule/schedule/models.py
@@ -52,7 +52,7 @@ class Tutor(models.Model):
     LastName = models.CharField(max_length=50)
     Email = models.EmailField(max_length=100)
     PhoneNumber = models.CharField(max_length=10)
-    RequestedHours = models.IntegerField()
+    RequestedHours = models.DecimalField(max_digits=3, decimal_places=1)
 
     def __str__(self):
         return str(self.FirstName + " " + self.LastName)
